% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ghuser.R
\name{check_gh_following}
\alias{check_gh_following}
\title{Check all SSH keys of GitHub users a particular account is following}
\usage{
check_gh_following(gh_user)
}
\arguments{
\item{gh_user}{a single GitHub user id}
}
\value{
data frame
}
\description{
Check all SSH keys of GitHub users a particular account is following
}
\note{
this may take a while for accounts that follow many users
}
\examples{
\dontrun{
check_gh_following("koenrh")
}
}
