% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-gh-org-members.R
\name{check_gh_org_members}
\alias{check_gh_org_members}
\title{Check all SSH keys of GitHub users that are members of a specified GitHub organization}
\usage{
check_gh_org_members(gh_org)
}
\arguments{
\item{gh_org}{a single GitHub organization name (e.g. "\code{GreyNoise-Intelligence}")}
}
\value{
data frame
}
\description{
Check all SSH keys of GitHub users that are members of a specified GitHub organization
}
\note{
this may take a while for organizations with many users
}
\examples{
\dontrun{
check_gh_org_members("GreyNoise-Intelligence")
}
}
