% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ghuser.R
\name{check_gh_repo_contributors}
\alias{check_gh_repo_contributors}
\title{Check all SSH keys of GitHub users a particular account is following}
\usage{
check_gh_repo_contributors(owner, repo)
}
\arguments{
\item{owner}{github owner name}

\item{repo}{github repo}
}
\value{
data frame
}
\description{
Check all SSH keys of GitHub users a particular account is following
}
\note{
this may take a while for repos with a large contributor list; this will also ignore "bots"
}
\examples{
\dontrun{
check_gh_repo_contributors("koenrh")
}
}
