% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ghuser.R
\name{check_gh_user_keys}
\alias{check_gh_user_keys}
\title{Check one or more GitHub user's keys}
\usage{
check_gh_user_keys(gh_users)
}
\arguments{
\item{gh_users}{(chr) a character vector of GitHub user ids}
}
\value{
data frame
}
\description{
Check one or more GitHub user's keys
}
\examples{
\dontrun{
check_gh_user_keys("hrbrmstr")
}
}
