% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ssh-key.R
\name{check_ssh_pub_key}
\alias{check_ssh_pub_key}
\title{Check one SSH public key}
\usage{
check_ssh_pub_key(keys)
}
\arguments{
\item{key}{Either a path to one or more files or literal public key strings}
}
\value{
data frame
}
\description{
Check one SSH public key
}
\examples{
\dontrun{
check_ssh_pub_key("~/.ssh/id_rsa.pub")
}
}
