% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubcheck-package.R
\docType{package}
\name{pubcheck}
\alias{pubcheck}
\alias{pubcheck-package}
\title{Check Safety of SSH Public Keys}
\description{
SSH is great! Poorly-configured SSH keys are not. Tools are provided to assess the
safety of SSH public keys in multiple contexts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/pubcheck}
  \item Report bugs at \url{https://github.com/hrbrmstr/pubcheck/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
