% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubcheck-package.R
\docType{package}
\name{pubcheck}
\alias{pubcheck}
\alias{pubcheck-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/pubcheck}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/pubcheck/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
